[csrng/dv] Add coverage for the transitions of the fips bit
This commit adds the coverage for the transitions of the fips bit at
the command input.
The transitions are recorded for each instance of the CSRNG and are
only recorded when a reseed command has completed. The alternative
would be to also track the transitions across instantiations as well,
but IMO it is more interesting to see the transition coverage during
single instantiations of the CSRNG.

Signed-off-by: Hakim Filali <hfilali@lowrisc.org>