[entropy_src/rtl] Absorb last word in HT window right after processing block
If the process_i signal goes high while SHA3 is processing a block, any word
that is left in the pipeline to be absorbed by SHA3 won't get absorbed in
the current health test window. This fix would allow SHA3 to absorb one
more word before starting to do the padding.

Signed-off-by: Hakim Filali <hfilali@lowrisc.org>