[otbnsim/oscar] Add initial tracking of secrets
This commit adds some additional tracking functionality
for secret sharings in the otbn simulation. The shares are
assigned a share id and a secret id along the location in dmem
whenever the secrets are loaded from dmem, the ids are traced in
the regs. This lays the groundwork for tracing shares throughout
the whole program. This would allow us to detect when styleguide
rules are violated. The locations of secrets in dmem are loaded
from a json config file. The same config file could potentially
be used to determine the rnd loads which generate new secrets.
Additional functionality would have to be added though.

Signed-off-by: h-filali <h.filali98@gmail.com>